reverting lightsaber version extracting classpath plath list into separate class
