+++ build.gradle
@@ -1,5 +1,5 @@
 allprojects {
-  version = '1.0.0-alpha14'
+  version = '1.0.0-alpha13'
   group = 'com.joom.lightsaber'
 
   ext.javaVersion = JavaVersion.VERSION_1_8
